Konverterer til less moduler, rydder i mappestruktur, og splitter opp kode
