+++ package.json
@@ -23,6 +23,7 @@
     "nav-frontend-ekspanderbartpanel-style": "^2.0.1",
     "nav-frontend-ikoner-assets": "^3.0.1",
     "nav-frontend-js-utils": "^1.0.19",
+    "nav-frontend-knapper-style": "^2.0.2",
     "nav-frontend-lenker": "^2.0.1",
     "nav-frontend-lenker-style": "^2.0.1",
     "nav-frontend-paneler-style": "^2.0.1",
@@ -30,6 +31,7 @@
     "nav-frontend-skjema-style": "^3.0.1",
     "nav-frontend-spinner": "^3.0.1",
     "nav-frontend-spinner-style": "^1.0.2",
+    "nav-frontend-tabell-style": "^2.0.2",
     "nav-frontend-typografi": "^4.0.1",
     "nav-frontend-typografi-style": "^2.0.1",
     "prop-types": "^15.7.2",
