+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TiltaksoversiktPage } from './page/tiltaksoversikt/TiltaksoversiktPage';
+import { TiltaksoversiktPage } from './component/page/tiltaksoversikt/TiltaksoversiktPage';
 
 export const App = () => {
 	return (
