+++ src/component/user-table/UserTable.module.less
@@ -1,0 +1,5 @@
+
+.ingenBrukere {
+  margin: 2rem auto 0 auto;
+  width: 400px;
+}
