+++ src/component/user-table/UserTable.tsx
@@ -2,34 +2,25 @@ import React from 'react';
 import { UserTableHeader } from './header/UserTableHeader';
 import { AlertStripeInfo } from 'nav-frontend-alertstriper';
 import { UserTableBody } from './body/UserTableBody';
-import { OrderByData } from './table-utils';
-import './UserTable.less';
-import { OrderByDirection, OrderByField } from './types';
 import { mockBrukere } from '../../mock/data/brukere';
+import styles from './UserTable.module.less'
+import "nav-frontend-tabell-style";
 
 export const UserTable = () => {
-	const orderByData: OrderByData = {
-		field: OrderByField.NAVN,
-		direction: OrderByDirection.ASC
-	};
 
-	const handleOnOrderByChanged = (nyData: OrderByData) => {
-		// setOrderByField(nyData.field);
-		// setOrderByDirection(nyData.direction);
-	};
+	const IngenBrukereAlertstripe = () =>
+		<AlertStripeInfo className={styles.ingenBrukere}>
+			<span role="alert" aria-live="polite">Fant ingen brukere</span>
+		</AlertStripeInfo>
 
-    return (
-    	<div role="table" aria-label="Brukere som har tiltak" aria-rowcount={mockBrukere.length} className="user-table">
-		    <UserTableHeader orderByData={orderByData} onOrderByChanged={handleOnOrderByChanged} />
-		    {
-			    mockBrukere.length === 0
-				    ? (
-					    <AlertStripeInfo className="user-table__no-users">
-						    <span role="alert" aria-live="polite">Fant ingen brukere</span>
-					    </AlertStripeInfo>
-			        )
-				    : <UserTableBody brukere={mockBrukere} />
-		    }
-	    </div>
+	return (
+    	<table className="tabell tabell--stripet">
+		    <UserTableHeader />
+			{
+				mockBrukere.length === 0
+					? <IngenBrukereAlertstripe/>
+					: <UserTableBody brukere={mockBrukere}/>
+			}
+		</table>
     );
 };
