--- src/component/user-table/body/UserTableBody.less
@@ -1,54 +1,0 @@
-@import "~nav-frontend-core/less/_variabler";
-@import "src/component/user-table/UserTable";
-
-.user-table-body {
-  padding: 0;
-  margin: 0;
-  border: 1px solid @navLysGra;
-
-  .user-table-row:nth-child(even) {
-    background-color: white;
-  }
-
-  .user-table-row:nth-child(odd) {
-    background-color: #f7f7f7;
-  }
-}
-
-.user-table-row {
-  display: block;
-
-  &__innhold {
-    display: grid;
-    .column-template();
-    justify-items: start;
-    padding: 1rem 2rem;
-    cursor: pointer;
-    text-align: center;
-    align-items: end;
-    color: @navMorkGra;
-    text-decoration: none;
-  }
-
-  &__innhold--maskert {
-    cursor: not-allowed;
-  }
-
-  &:hover {
-    position: relative;
-    z-index: 1;
-
-    box-shadow: 0 5px 5px fade(@navGra20, 60%), 0 -5px 5px fade(@navGra20, 60%);
-  }
-
-  .status {
-    display: grid;
-    grid-template-columns: 2rem auto;
-    align-items: end;
-
-    .status_ikon {
-      max-width: 26px;
-    }
-  }
-
-}
