--- src/component/user-table/header/UserTableHeader.less
@@ -1,32 +1,0 @@
-@import "~nav-frontend-core/less/_variabler";
-@import "src/component/user-table/UserTable";
-
-.user-table-header {
-  display: grid;
-  .column-template();
-  justify-items: start;
-  padding: 0 2rem 1rem 2rem;
-  border-bottom: 2px solid @navGra60;
-
-  .table-header-field {
-    font-size: 1rem;
-    font-weight: bolder;
-    color: @navBla;
-    padding: 0;
-
-    background: transparent;
-    border: none;
-    cursor: pointer;
-    white-space: nowrap;
-  }
-
-  .table-header-field__order-icon {
-    width: 16px;
-    height: 16px;
-    margin-left: 0.25rem;
-  }
-
-  .table-header-field__order-icon--asc {
-    transform: rotate(180deg);
-  }
-}
