+++ src/component/user-table/pagination-bar/PaginationBar.module.less
@@ -1,0 +1,11 @@
+.paginationBar {
+  display: flex;
+  justify-content: space-between;
+  padding: 0 2rem 2rem 2rem;
+}
+
+.paginering {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
