+++ src/component/user-table/pagination-bar/PagineringKnapp.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import cls from 'classnames';
+
+export const PagineringKnapp = (props: React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement>) => {
+    const { className, children, ...rest } = props;
+    return (
+        <button className={cls(className, 'paginering__knapp')} {...rest}>
+            {props.children}
+        </button>
+    );
+}
