+++ src/index.less
@@ -5,5 +5,5 @@
 
 html, body {
   margin: 0;
-  background-color: #f4f4f4; // TODO: skal vi heller bruke en farge i designsystemet?
+  background-color: @navGraBakgrunn;
 }
